feat(auth): extract web specific components
move AuthTokenUtils to services/ since it closely resembles a service
create simple key-value store, implement using cookies for web
use web specific implementation only on web

this fixes widget tests
