+++ lib/services/web/key_value_store_web_stub.dart
@@ -1,0 +1,4 @@
+
+import '../key_value_store.dart';
+
+KeyValueStore getKeyValueStoreWeb() => throw UnimplementedError();
